+++ src/web/assets/src/vue/export/organisms/forms/FormExport.vue
@@ -20,69 +20,94 @@
         <!--input type="hidden" name="start" :value="start">
         <input type="hidden" name="end" :value="end"-->
 
-        <div class="grid grid-cols-5">
-            <label class="pr-4">
-                <span class="text-xs font-bold text-gray-400 block mb-1">Which schools?</span>
-                <select name="school" class="block h-10 px-1 rounded-md bg-gray-100 w-full border-none">
-                    <option default value="all">All schools</option>
-                    <option default value="prior">Priority schools</option>
-                </select>
-            </label>
-            <label class="pr-4">
-                <span class="text-xs font-bold text-gray-400 block mb-1">Select an export type</span>
-                <select name="type" class="block h-10 px-1 rounded-md bg-gray-100 w-full border-none">
-                    <option default value="attendee">Attendee Data</option>
-                    <option default value="event">Event ID reference table</option>
-                    <option default value="subscriptions">Mailing list subscribers</option>
-                    <option default value="school-attendee">School data</option>
-                </select>
-            </label>
-            <label class="col-span-2 pr-4">
-                <span class="text-xs font-bold text-gray-400 block mb-1">Time period</span>
-                <div class="flex w-full">
-                    <div class="inline-block flex-grow">
+        <div class="grid grid-cols-3 gap-4">
+
+            <div>
+
+                <label>
+                    <span class="text-xs font-bold text-gray-400 block mb-1">Select an export type</span>
+                    <select name="exportType" class="block h-10 px-1 rounded-md bg-gray-100 w-full border-none mb-2">
+                        <option default value="attendee">Attendee Data</option>
+                        <option default value="event">Event ID reference table</option>
+                        <option default value="subscriptions">Mailing list subscribers</option>
+                        <option default value="school-attendee">School data</option>
+                    </select>
+                </label>
+
+            </div>
+
+            <div>
+                <label>
+                    <span class="text-xs font-bold text-gray-400 block mb-1">Select an event type</span>
+                    <select name="eventType" class="block h-10 px-1 rounded-md bg-gray-100 w-full border-none mb-2">
+                        <option default value="16,17,25">All events</option>
+                        <option default value="17">Location Event</option>
+                        <option default value="16">Online Event</option>
+                        <option default value="25">Hybrid Event</option>
+                    </select>
+                </label>
+            </div>
+
+            <div>
+                <label>
+                    <span class="text-xs font-bold text-gray-400 block mb-1">Which schools?</span>
+                    <select name="school" class="block h-10 px-1 rounded-md bg-gray-100 w-full border-none mb-2">
+                        <option default value="all">All schools</option>
+                        <option default value="prior">Priority schools</option>
+                    </select>
+                </label>
+            </div>
+
+            <div class="col-span-2">
+                <label>
+                    <span class="text-xs font-bold text-gray-400 block mb-1">Time period</span>
+                    <div class="w-full grid grid-cols-2 gap-4">
                         <div class="flex">
-                            <span class="text-xs text-gray-400 inline-block">From</span>
-                            <input type="date" v-model="startDate" name="start" :max="yesterday" :class="[
-                                'px-2 flex-grow py-2 text-sm text-gray-600 appearance-none box-border bg-gray-100 rounded-lg border-solid',
-                                errors?.startDate ? 'border-red-300' : 'border-gray-100'
-                            ]"/>
+                            <div class="flex w-full">
+                                <span class="text-xs text-gray-400 inline-block">From</span>
+                                <input type="date" v-model="startDate" name="start" :max="yesterday" :class="[
+                                    'px-2 flex-grow py-2 text-sm text-gray-600 appearance-none box-border bg-gray-100 rounded-lg border-solid',
+                                    errors?.startDate ? 'border-red-300' : 'border-gray-100'
+                                ]"/>
+                            </div>
+                            <span v-if="errors?.startDate" class="text-xs p-2 inline-block font-bold text-red-300">{{errors?.startDate}}</span>
                         </div>
-                        <span v-if="errors?.startDate" class="text-xs p-2 inline-block font-bold text-red-300">{{errors?.startDate}}</span>
-                    </div>
-                    <div class="inline-block flex-grow">
                         <div class="flex">
-                            <span class="text-xs text-gray-400 inline-block">To</span>
-                            <input type="date" v-model="endDate" name="end" :max="today" :class="[
-                                'flex-grow px-2 py-2 text-sm text-gray-600 appearance-none box-border bg-gray-100 rounded-lg border-solid',
-                                errors?.endDate ? 'border-red-300' : 'border-gray-100'
-                            ]"/>
+                            <div class="flex w-full">
+                                <span class="text-xs text-gray-400 inline-block">To</span>
+                                <input type="date" v-model="endDate" name="end" :max="today" :class="[
+                                    'flex-grow px-2 py-2 text-sm text-gray-600 appearance-none box-border bg-gray-100 rounded-lg border-solid',
+                                    errors?.endDate ? 'border-red-300' : 'border-gray-100'
+                                ]"/>
+                            </div>
+                            <span v-if="errors?.endDate" class="text-xs p-2 inline-block font-bold text-red-300">{{errors?.endDate}}</span>
                         </div>
-                        <span v-if="errors?.endDate" class="text-xs p-2 inline-block font-bold text-red-300">{{errors?.endDate}}</span>
                     </div>
-                </div>
-                <p class="mt-0 text-xs">Predefined selections:
-                    <span class="text-xs text-blue-600 cursor-pointer" @click.prevent="setPredefined('3m')">Last&nbsp;3&nbsp;months</span>,
-                    <span class="text-xs text-blue-600 cursor-pointer" @click.prevent="setPredefined('6m')">Last&nbsp;6&nbsp;months</span>,
-                    <span class="text-xs text-blue-600 cursor-pointer" @click.prevent="setPredefined('12m')">Last&nbsp;year</span>,
-                    <span class="text-xs text-blue-600 cursor-pointer" @click.prevent="setPredefined('cay')">Current&nbsp;academic&nbsp;year</span> or
-                    <span class="text-xs text-blue-600 cursor-pointer" @click.prevent="setPredefined('lay')">Last&nbsp;academic&nbsp;year</span>
-                </p>
-            </label>
-            <label>
-                <span class="text-xs font-bold text-gray-400 block mb-1">Filter by tag</span>
-                <!--input v-model="tag" name="tag" :class="[
-                    'px-2 w-full py-2 text-sm text-gray-600 appearance-none box-border bg-gray-100 rounded-lg border-solid',
-                    errors?.startDate ? 'border-red-300' : 'border-gray-100'
-                ]"/-->
-                <input-tags />
-            </label>
+                    <p class="-mt-4 text-xs">Predefined selections:
+                        <span class="text-xs text-blue-600 cursor-pointer" @click.prevent="setPredefined('3m')">Last&nbsp;3&nbsp;months</span>,
+                        <span class="text-xs text-blue-600 cursor-pointer" @click.prevent="setPredefined('6m')">Last&nbsp;6&nbsp;months</span>,
+                        <span class="text-xs text-blue-600 cursor-pointer" @click.prevent="setPredefined('12m')">Last&nbsp;year</span>,
+                        <span class="text-xs text-blue-600 cursor-pointer" @click.prevent="setPredefined('cay')">Current&nbsp;academic&nbsp;year</span> or
+                        <span class="text-xs text-blue-600 cursor-pointer" @click.prevent="setPredefined('lay')">Last&nbsp;academic&nbsp;year</span>
+                    </p>
+                </label>
+            </div>
+            <div>
+                <label>
+                    <span class="text-xs font-bold text-gray-400 block mb-1">Filter by tag</span>
+                    <!--input v-model="tag" name="tag" :class="[
+                        'px-2 w-full py-2 text-sm text-gray-600 appearance-none box-border bg-gray-100 rounded-lg border-solid',
+                        errors?.startDate ? 'border-red-300' : 'border-gray-100'
+                    ]"/-->
+                    <input-tags />
+                </label>
+            </div>
         </div>
     </form>
 </template>
 
 <script lang="ts">
-    import {defineComponent, ref } from 'vue'
+    import { defineComponent, ref } from 'vue'
     import InputTag from "@/vue/export/atoms/inputs/InputTag.vue";
 
     export default defineComponent({
@@ -99,9 +124,7 @@
                 required: true
             }
         },
-        setup(){
-            const start = ref(null)
-            const end = ref(null)
+        setup() {
             const startDate = ref('2017-07-04')
             const endDate = ref(
                 new Date().getFullYear()+'-'+